requirements.txt
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==5.2.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pillow==12.1.1
pyngrok==7.5.1
python-telegram-bot==22.6
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
